requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.5.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-multipart==0.0.6
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
requests==2.31.0
email-validator==2.1.0